Split `BuildScriptCompileAvoidanceIntegrationTest`
