+++ subprojects/kotlin-dsl/src/integTest/kotlin/org/gradle/kotlin/dsl/compile/BuildScriptCompileAvoidanceIntegrationTest.kt
@@ -1,39 +1,28 @@
-package org.gradle.kotlin.dsl.integration
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import org.gradle.integtests.fixtures.BuildOperationsFixture
+package org.gradle.kotlin.dsl.compile
+
 import org.gradle.integtests.fixtures.UnsupportedWithConfigurationCache
-import org.gradle.kotlin.dsl.fixtures.AbstractKotlinIntegrationTest
 import org.gradle.kotlin.dsl.provider.KOTLIN_SCRIPT_COMPILATION_AVOIDANCE_ENABLED_PROPERTY
-import org.gradle.util.Matchers.isEmpty
-import org.hamcrest.CoreMatchers.containsString
-import org.hamcrest.CoreMatchers.endsWith
-import org.hamcrest.CoreMatchers.hasItem
-import org.hamcrest.MatcherAssert.assertThat
-import org.hamcrest.Matchers.hasSize
 import org.junit.Assert.assertTrue
-import org.junit.Before
 import org.junit.Test
-import java.io.File
-import java.util.UUID
-import java.util.regex.Pattern
 
 
-class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest() {
-
-    companion object CacheBuster {
-        var cacheBuster = UUID.randomUUID()
-    }
-
-    @Before
-    fun init() {
-        cacheBuster = UUID.randomUUID()
-
-        withSettings(
-            """
-            rootProject.name = "test-project"
-            """
-        )
-    }
+class BuildScriptCompileAvoidanceIntegrationTest : AbstractCompileAvoidanceIntegrationTest() {
 
     @Test
     fun `script compilation avoidance can be disabled via a system property`() {
@@ -248,122 +237,6 @@ class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest
     }
 
     @Test
-    fun `avoids buildscript recompilation on non ABI change in precompiled script plugin`() {
-        val pluginId = "my-plugin"
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                println("foo")
-            """
-        )
-        withUniqueScript(
-            """
-                plugins {
-                    id("$pluginId")
-                }
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                println("bar")
-            """
-        )
-        configureProject().assertBuildScriptCompilationAvoided().assertOutputContains("bar")
-    }
-
-    @Test
-    fun `recompiles buildscript when new task is registered in precompiled script plugin`() {
-        val pluginId = "my-plugin"
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                println("foo")
-            """
-        )
-        withUniqueScript(
-            """
-                plugins {
-                    id("$pluginId")
-                }
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                println("bar")
-                tasks.register("foo")
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("bar")
-    }
-
-    @Test
-    fun `avoids buildscript recompilation when task is configured in precompiled script plugin`() {
-        val pluginId = "my-plugin"
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                println("foo")
-                tasks.register("foo")
-            """
-        )
-        withUniqueScript(
-            """
-                plugins {
-                    id("$pluginId")
-                }
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                tasks.register("foo") { doLast { println("bar from task") } }
-            """
-        )
-        configureProject("foo").assertBuildScriptCompilationAvoided().assertOutputContains("bar from task")
-    }
-
-    @Test
-    fun `recompiles buildscript when plugins applied from a precompiled plugin change`() {
-        val pluginId = "my-plugin"
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                plugins {
-                    id("java-library")
-                }
-                println("foo")
-            """
-        )
-        withUniqueScript(
-            """
-                plugins {
-                    id("$pluginId")
-                }
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                plugins {
-                    id("java")
-                }
-                println("bar")
-            """
-        )
-        configureProject().assertBuildScriptCompiled().assertOutputContains("bar")
-    }
-
-    @Test
     fun `recompiles buildscript on inline function change in buildSrc class`() {
         val className = givenKotlinClassInBuildSrcContains(
             """
@@ -436,56 +309,6 @@ class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest
     }
 
     @Test
-    fun `avoids buildscript recompilation on public function change in buildSrc script`() {
-        val packageName = givenKotlinScriptInBuildSrcContains(
-            "Foo",
-            """
-            fun foo() {
-                println("foo")
-            }
-            """
-        )
-        withUniqueScript("$packageName.foo()")
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        givenKotlinScriptInBuildSrcContains(
-            "Foo",
-            """
-            fun foo() {
-                println("bar")
-            }
-            """
-        )
-        configureProject().assertBuildScriptCompilationAvoided().assertOutputContains("bar")
-    }
-
-    @Test
-    fun `recompiles buildscript on inline function change in buildSrc script`() {
-        val packageName = givenKotlinScriptInBuildSrcContains(
-            "Foo",
-            """
-            inline fun foo() {
-                println("foo")
-            }
-            """
-        )
-        withUniqueScript("$packageName.foo()")
-        configureProjectAndExpectCompileAvoidanceWarnings().assertBuildScriptCompiled().assertOutputContains("foo")
-            .assertContainsCompileAvoidanceWarning("buildSrc.jar: class com/example/FooKt: inline fun foo(): compile avoidance is not supported with public inline functions")
-
-        givenKotlinScriptInBuildSrcContains(
-            "Foo",
-            """
-            inline fun foo() {
-                println("bar")
-            }
-            """
-        )
-        configureProjectAndExpectCompileAvoidanceWarnings().assertBuildScriptBodyRecompiled().assertOutputContains("bar")
-            .assertContainsCompileAvoidanceWarning("buildSrc.jar: class com/example/FooKt: inline fun foo(): compile avoidance is not supported with public inline functions")
-    }
-
-    @Test
     fun `avoids buildscript recompilation on internal inline function change in buildSrc class`() {
         val className = givenKotlinClassInBuildSrcContains(
             """
@@ -507,63 +330,6 @@ class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest
             """
         )
         configureProject().assertBuildScriptCompilationAvoided().assertOutputContains("bar")
-    }
-
-    @Test
-    fun `recompiles buildscript on const val field change in buildSrc script`() {
-        val packageName = givenKotlinScriptInBuildSrcContains(
-            "foo",
-            """
-            const val FOO = "foo"
-            """
-        )
-        withUniqueScript("println($packageName.FOO)")
-        configureProject().assertBuildScriptCompiled().assertOutputContains("foo")
-
-        givenKotlinScriptInBuildSrcContains(
-            "foo",
-            """
-            const val FOO = "bar"
-            """
-        )
-        configureProject().assertBuildScriptBodyRecompiled().assertOutputContains("bar")
-    }
-
-    @Test
-    fun `recompiles buildscript when plugin extension registration name changes from a precompiled plugin`() {
-        val pluginId = "my-plugin"
-        val extensionClass = """
-            open class TestExtension {
-                var message = "some-message"
-            }
-        """
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                $extensionClass
-                project.extensions.create<TestExtension>("foo")
-            """
-        )
-        withUniqueScript(
-            """
-                plugins {
-                    id("$pluginId")
-                }
-                foo {
-                    message = "foo"
-                }
-            """
-        )
-        configureProject().assertBuildScriptCompiled()
-
-        withPrecompiledScriptPluginInBuildSrc(
-            pluginId,
-            """
-                $extensionClass
-                project.extensions.create<TestExtension>("bar")
-            """
-        )
-        configureProjectAndExpectCompileFailure("Unresolved reference: foo")
     }
 
     @Test
@@ -733,34 +499,6 @@ class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest
     }
 
     private
-    fun withUniqueScript(script: String) = withBuildScript(script).apply {
-        bustScriptCache()
-    }
-
-    private
-    fun withKotlinDslPluginInBuildSrc() {
-        // this is to force buildSrc/build.gradle.kts to be written to test-local buildscript cache
-        // and not to be reused from daemon's cache from other tests when daemon is in use
-        withBuildScriptIn("buildSrc", scriptWithKotlinDslPlugin())
-            .bustScriptCache()
-    }
-
-    private
-    fun File.bustScriptCache() {
-        appendText(
-            """
-                val cacheBuster = "$cacheBuster"
-            """
-        )
-    }
-
-    private
-    fun withPrecompiledScriptPluginInBuildSrc(pluginId: String, pluginSource: String) {
-        withKotlinDslPluginInBuildSrc()
-        withFile("buildSrc/src/main/kotlin/$pluginId.gradle.kts", pluginSource)
-    }
-
-    private
     fun buildJarForBuildScriptClasspath(classBody: String): Pair<String, String> {
         val baseDir = "buildscript"
         withSettingsIn(
@@ -810,149 +548,5 @@ class BuildScriptCompileAvoidanceIntegrationTest : AbstractKotlinIntegrationTest
         val jarPath = "$baseDir/build/libs/buildscript.jar"
         assertTrue(existing(jarPath).exists())
         return Pair(className, jarPath)
-    }
-
-    private
-    fun givenJavaClassInBuildSrcContains(classBody: String, classAnnotations: String = ""): String =
-        javaSourceFile("buildSrc", classBody, classAnnotations)
-
-    private
-    fun givenKotlinClassInBuildSrcContains(classBody: String): String {
-        withKotlinDslPluginInBuildSrc()
-        return kotlinClassSourceFile("buildSrc", classBody)
-    }
-
-    private
-    fun givenKotlinScriptInBuildSrcContains(scriptName: String, scriptBody: String, scriptPrefix: String = ""): String {
-        withKotlinDslPluginInBuildSrc()
-        return kotlinScriptSourceFile("buildSrc", scriptName, scriptBody, scriptPrefix)
-    }
-
-    private
-    fun javaSourceFile(baseDir: String, classBody: String, classAnnotations: String = ""): String {
-        val className = "Foo"
-        withFile(
-            "$baseDir/src/main/java/com/example/$className.java",
-            """
-            package com.example;
-            $classAnnotations
-            public class $className {
-                $classBody
-            }
-            """
-        )
-        return "com.example.$className"
-    }
-
-    private
-    fun kotlinClassSourceFile(baseDir: String, classBody: String): String {
-        val className = "Foo"
-        val packageName = kotlinScriptSourceFile(
-            baseDir,
-            className,
-            """
-            class $className {
-                $classBody
-            }
-            """
-        )
-        return "$packageName.$className"
-    }
-
-    private
-    fun kotlinScriptSourceFile(baseDir: String, scriptName: String, scriptBody: String, scriptPrefix: String = ""): String {
-        withFile(
-            "$baseDir/src/main/kotlin/com/example/$scriptName.kt",
-            """
-            $scriptPrefix
-            package com.example
-            $scriptBody
-            """
-        )
-        return "com.example"
-    }
-
-    private
-    fun configureProject(vararg tasks: String): BuildOperationsAssertions {
-        val buildOperations = BuildOperationsFixture(executer, testDirectoryProvider)
-        val output = executer.withTasks(*tasks).run().normalizedOutput
-        return BuildOperationsAssertions(buildOperations, output)
-    }
-
-    private
-    fun configureProjectAndExpectCompileAvoidanceWarnings(vararg tasks: String): BuildOperationsAssertions {
-        val buildOperations = BuildOperationsFixture(executer, testDirectoryProvider)
-        val output = executer.withArgument("--info").withTasks(*tasks).run().normalizedOutput
-        return BuildOperationsAssertions(buildOperations, output, true)
-    }
-
-    private
-    fun configureProjectAndExpectCompileFailure(expectedFailure: String) {
-        val error = executer.runWithFailure().error
-        assertThat(error, containsString(expectedFailure))
-    }
-}
-
-
-private
-class BuildOperationsAssertions(buildOperationsFixture: BuildOperationsFixture, val output: String, val expectWarnings: Boolean = false) {
-    private
-    val classpathCompileOperations = buildOperationsFixture.all(Pattern.compile("Compile script build.gradle.kts \\(CLASSPATH\\)"))
-
-    private
-    val bodyCompileOperations = buildOperationsFixture.all(Pattern.compile("Compile script build.gradle.kts \\(BODY\\)"))
-
-    private
-    val compileAvoidanceWarnings = output.lines()
-        .filter { it.startsWith("Cannot use Kotlin build script compile avoidance with") }
-        // filter out avoidance warnings for versioned jars - those come from Kotlin/libraries that don't change when code under test changes
-        .filterNot { it.contains(Regex("\\d.jar: ")) }
-
-    init {
-        if (!expectWarnings) {
-            assertThat(compileAvoidanceWarnings, isEmpty())
-        }
-    }
-
-    fun assertBuildScriptCompiled(): BuildOperationsAssertions {
-        if (classpathCompileOperations.isNotEmpty() || bodyCompileOperations.isNotEmpty()) {
-            return this
-        }
-        throw AssertionError("Expected script to be compiled, but it wasn't.")
-    }
-
-    fun assertBuildScriptBodyRecompiled(): BuildOperationsAssertions {
-        if (bodyCompileOperations.size == 1) {
-            return this
-        }
-        if (bodyCompileOperations.isEmpty()) {
-            throw AssertionError("Expected build script body to be recompiled, but it wasn't.")
-        }
-        throw AssertionError("Expected build script body to be recompiled, but there was more than one body compile operation: $bodyCompileOperations")
-    }
-
-    fun assertBuildScriptCompilationAvoided(): BuildOperationsAssertions {
-        if (classpathCompileOperations.isEmpty() && bodyCompileOperations.isEmpty()) {
-            return this
-        }
-        throw AssertionError(
-            "Expected script compilation to be avoided, but the buildscript was recompiled. " +
-                "classpath compile operations: $classpathCompileOperations, body compile operations: $bodyCompileOperations"
-        )
-    }
-
-    fun assertOutputContains(expectedOutput: String): BuildOperationsAssertions {
-        assertThat(output, containsString(expectedOutput))
-        return this
-    }
-
-    fun assertContainsCompileAvoidanceWarning(end: String): BuildOperationsAssertions {
-        assertThat(compileAvoidanceWarnings, hasItem(endsWith(end)))
-        return this
-    }
-
-    fun assertNumberOfCompileAvoidanceWarnings(n: Int): BuildOperationsAssertions {
-        assertThat(compileAvoidanceWarnings, hasSize(n))
-        return this
     }
 }
